+++ examples/main.rs
@@ -5,9 +5,17 @@
 // Copyright (c) DUSK NETWORK. All rights reserved.
 
 use clap::{App, Arg};
+use kadcast::config::Config;
 use kadcast::{MessageInfo, NetworkListen, Peer};
 use rustc_tools_util::{get_version_info, VersionInfo};
 use std::io::{self, BufRead};
+
+use serde_derive::Deserialize;
+use serde_derive::Serialize;
+#[derive(Serialize, Deserialize)]
+struct General {
+    kadcast: kadcast::config::Config,
+}
 
 #[tokio::main]
 pub async fn main() {
@@ -76,25 +84,18 @@ pub async fn main() {
     tracing::subscriber::set_global_default(subscriber)
         .expect("Failed on subscribe tracing");
 
-    let public_address =
+    let mut conf = Config::default();
+    conf.public_address =
         matches.value_of("public_address").unwrap().to_string();
-
-    let listen_address =
+    conf.listen_address =
         matches.value_of("listen_address").map(|a| a.to_string());
-
-    let bootstrapping_nodes = matches
+    conf.bootstrapping_nodes = matches
         .values_of("bootstrap")
         .unwrap_or_default()
         .map(|s| s.to_string())
         .collect();
 
-    let mut builder =
-        Peer::builder(public_address, bootstrapping_nodes, DummyListener {})
-            .with_listen_address(listen_address);
-    builder
-        .transport_conf()
-        .extend(kadcast::transport::default_configuration());
-    let peer = builder.build();
+    let peer = Peer::new(conf, DummyListener {});
     loop {
         let stdin = io::stdin();
         for message in stdin.lock().lines().flatten() {