+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Added
+
+- Add `kadcast::Config` [#96]
+
+### Removed
+
+- `PeerBuilder` in favor of `Peer::new()`
+
 ## [0.3.0] - 07-01-22
 ### Added
 
@@ -67,3 +77,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [#87]: https://github.com/dusk-network/kadcast/issues/87
 [#90]: https://github.com/dusk-network/kadcast/issues/90
 [#92]: https://github.com/dusk-network/kadcast/issues/92
+[#96]: https://github.com/dusk-network/kadcast/issues/96